fix: allow factories in seeds
This also splits out test helpers into individual files. Factories weren't building properly due to a Typescript error, so short term we need to avoid importing the graphql server from pages/api/graphql.
