+++ packages/create-bison-app/template/prisma/seed.ts
@@ -1,13 +1,9 @@
-import { PrismaClient } from '@prisma/client';
-const prisma = new PrismaClient();
+// import { UserFactory } from '../tests/factories';
+// import { Role } from '../types';
+// import { prisma } from '../lib/prisma';
 
-async function main() {
-  // TODO: seeds
-  console.log('no seeds yet!');
+export async function seed() {
+  // Add seeds here. You can use factories or raw prisma.create/upsert calls.
+  // await UserFactory.create({ email: 'hello@wee.net' });
+  // await UserFactory.create({ roles: [Role.ADMIN] });
 }
-
-main()
-  .catch((e) => console.error(e))
-  .finally(async () => {
-    await prisma.$disconnect();
-  });
